# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ComIT Viribus API
  description: REST API для SPA (без эндпоинтов ИИ-чата)
  version: 1.0.0
paths:
...
  /api/library/interests:
    post:
      tags:
      - analytics
      summary: Post Library Interests
      operationId: post_library_interests_api_library_interests_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestsPayload'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/library/showcase:
    post:
      tags:
      - library-admin
      summary: Admin Create Showcase
      operationId: admin_create_showcase_api_admin_library_showcase_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseCreateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryShowcaseItem'
        '409':
          description: Уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/admin/library/showcase/{item_id}:
    patch:
      tags:
      - library-admin
      summary: Admin Update Showcase
      operationId: admin_update_showcase_api_admin_library_showcase__item_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowcaseUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryShowcaseItem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - library-admin
      summary: Admin Delete Showcase
      operationId: admin_delete_showcase_api_admin_library_showcase__item_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/library/interests:
    post:
      tags:
      - library-admin
      summary: Admin Create Interest
      operationId: admin_create_interest_api_admin_library_interests_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestOptionCreateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestOption'
        '409':
          description: Уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/admin/library/interests/{interest_id}:
    patch:
      tags:
      - library-admin
      summary: Admin Update Interest
      operationId: admin_update_interest_api_admin_library_interests__interest_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: interest_id
        in: path
        required: true
        schema:
          type: string
          title: Interest Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestOptionUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestOption'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - library-admin
      summary: Admin Delete Interest
      operationId: admin_delete_interest_api_admin_library_interests__interest_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: interest_id
        in: path
        required: true
        schema:
          type: string
          title: Interest Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/library/tags:
    post:
      tags:
      - library-admin
      summary: Admin Create Tag
      operationId: admin_create_tag_api_admin_library_tags_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryTagCreateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleTag'
        '409':
          description: Уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/admin/library/tags/{tag_id}:
    patch:
      tags:
      - library-admin
      summary: Admin Update Tag
      operationId: admin_update_tag_api_admin_library_tags__tag_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          title: Tag Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryTagUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleTag'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - library-admin
      summary: Admin Delete Tag
      operationId: admin_delete_tag_api_admin_library_tags__tag_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          title: Tag Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ArticleTag:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        tone:
          type: string
          title: Tone
      type: object
      required:
      - id
      - label
      - tone
      title: ArticleTag
...
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorDetail
      description: Тело ответа HTTPException (например 404).
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InterestOption:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        selected:
          type: boolean
          title: Selected
      type: object
      required:
      - id
      - label
      - selected
      title: InterestOption
    InterestOptionCreateBody:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        selected:
          type: boolean
          title: Selected
          default: false
      type: object
      required:
      - id
      - label
      title: InterestOptionCreateBody
    InterestOptionUpdateBody:
      properties:
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        selected:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Selected
        sortOrder:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sortorder
      type: object
      title: InterestOptionUpdateBody
    InterestsPayload:
      properties:
        interests:
          items:
            type: string
          type: array
          title: Interests
        ts:
          type: integer
          title: Ts
      type: object
      required:
      - interests
      - ts
      title: InterestsPayload
...
    LibraryHero:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        updatedLabel:
          type: string
          title: Updatedlabel
        providerLabel:
          type: string
          title: Providerlabel
        description:
          type: string
          title: Description
      type: object
      required:
      - id
      - title
      - updatedLabel
      - providerLabel
      - description
      title: LibraryHero
    LibraryShowcaseItem:
      properties:
        id:
          type: string
          title: Id
        brandLabel:
          type: string
          title: Brandlabel
        eyebrow:
          type: string
          title: Eyebrow
        title:
          type: string
          title: Title
        imageUrl:
          type: string
          title: Imageurl
        hero:
          $ref: '#/components/schemas/LibraryHero'
      type: object
      required:
      - id
      - brandLabel
      - eyebrow
      - title
      - imageUrl
      - hero
      title: LibraryShowcaseItem
    LibraryTagCreateBody:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        tone:
          type: string
          title: Tone
        sortOrder:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sortorder
      type: object
      required:
      - id
      - label
      - tone
      title: LibraryTagCreateBody
    LibraryTagUpdateBody:
      properties:
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
        tone:
          anyOf:
          - type: string
          - type: 'null'
          title: Tone
        sortOrder:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sortorder
      type: object
      title: LibraryTagUpdateBody
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    ShowcaseCreateBody:
      properties:
        id:
          type: string
          title: Id
        brandLabel:
          type: string
          title: Brandlabel
        eyebrow:
          type: string
          title: Eyebrow
        title:
          type: string
          title: Title
        imageUrl:
          type: string
          title: Imageurl
        hero:
          $ref: '#/components/schemas/LibraryHero'
      type: object
      required:
      - id
      - brandLabel
      - eyebrow
      - title
      - imageUrl
      - hero
      title: ShowcaseCreateBody
    ShowcaseUpdateBody:
      properties:
        brandLabel:
          anyOf:
          - type: string
          - type: 'null'
          title: Brandlabel
        eyebrow:
          anyOf:
          - type: string
          - type: 'null'
          title: Eyebrow
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        imageUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Imageurl
        hero:
          anyOf:
          - $ref: '#/components/schemas/LibraryHero'
          - type: 'null'
        sortOrder:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sortorder
      type: object
      title: ShowcaseUpdateBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/jwt/login